+++ ca/django_ca/managers.py
@@ -21,6 +21,7 @@ from cryptography import x509
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric import dsa
+from cryptography.hazmat.primitives.asymmetric import ec
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.primitives.serialization import Encoding
 from cryptography.hazmat.primitives.serialization import PrivateFormat
@@ -45,6 +46,7 @@ from .utils import TLS_FEATURE_MAPPING
 from .utils import get_cert_builder
 from .utils import is_power2
 from .utils import parse_general_name
+from .utils import parse_key_curve
 
 
 class CertificateManagerMixin(object):
@@ -72,19 +74,18 @@ class CertificateManagerMixin(object):
 
 
 class CertificateAuthorityManager(CertificateManagerMixin, models.Manager):
-    def init(self, name, key_size, key_type, algorithm, expires, parent, subject, pathlen=None,
+    def init(self, name, key_type, algorithm, expires, parent, subject, pathlen=None,
              issuer_url=None, issuer_alt_name=None, crl_url=None, ocsp_url=None,
              ca_issuer_url=None, ca_crl_url=None, ca_ocsp_url=None, name_constraints=None,
-             password=None, parent_password=None):
+             password=None, parent_password=None, ecc_curve=None, key_size=None):
         """Create a new certificate authority.
 
         Parameters
         ----------
 
-        key_size : int
-            Integer, must be a power of two (e.g. 2048, 4096, ...)
         key_type: str, optional
-            Either ``"RSA"`` or ``"DSA"`` for a RSA or DSA key, with ``"RSA"`` being the default.
+            The type of private key to generate, must be one of ``"RSA"``, ``"DSA"`` or ``"ECC"``, with
+            ``"RSA"`` being the default.
         algorithm : :py:class:`~cryptography:cryptography.hazmat.primitives.hashes.HashAlgorithm`
             Hash algorithm used when signing the certificate. Must be an instance of
             :py:class:`~cryptography:cryptography.hazmat.primitives.hashes.HashAlgorithm`, e.g.
@@ -101,14 +102,21 @@ class CertificateAuthorityManager(CertificateManagerMixin, models.Manager):
             Password to encrypt the private key with.
         parent_password : bytes, optional
             Password that the private key of the parent CA is encrypted with.
+        ecc_curve : str or EllipticCurve, optional
+            The elliptic curve to use for ECC type keys, passed verbatim to
+            :py:func:`~django_ca.utils.parse_key_curve`.
+        key_size : int, optional
+            Integer specifying the key size, must be a power of two (e.g. 2048, 4096, ...) unused if
+            ``key_type="ECC"`` but required otherwise.
         """
         # NOTE: This is already verified by KeySizeAction, so none of these checks should ever be
         #       True in the real world. None the less they are here as a safety precaution.
-        if not is_power2(key_size):
-            raise RuntimeError("%s: Key size must be a power of two." % key_size)
-        elif key_size < ca_settings.CA_MIN_KEY_SIZE:
-            raise RuntimeError("%s: Key size must be least %s bits."
-                               % (key_size, ca_settings.CA_MIN_KEY_SIZE))
+        if key_type != 'ECC':
+            if not is_power2(key_size):
+                raise RuntimeError("%s: Key size must be a power of two." % key_size)
+            elif key_size < ca_settings.CA_MIN_KEY_SIZE:
+                raise RuntimeError("%s: Key size must be least %s bits."
+                                   % (key_size, ca_settings.CA_MIN_KEY_SIZE))
 
         pre_create_ca.send(
             sender=self.model, name=name, key_size=key_size, key_type=key_type, algorithm=algorithm,
@@ -119,6 +127,9 @@ class CertificateAuthorityManager(CertificateManagerMixin, models.Manager):
 
         if key_type == 'DSA':
             private_key = dsa.generate_private_key(key_size=key_size, backend=default_backend())
+        elif key_type == 'ECC':
+            ecc_curve = parse_key_curve(ecc_curve)
+            private_key = ec.generate_private_key(ecc_curve, default_backend())
         else:
             private_key = rsa.generate_private_key(public_exponent=65537, key_size=key_size,
                                                    backend=default_backend())